# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Flora API",
    "version": "1.0.0",
    "description": "REST API for managing a bouquets collection."
  },
  "servers": [
    { "url": "/api", "description": "API base path" }
  ],
...
  "paths": {
...
    "/bouquets/{id}": {
      "get": {
        "tags": ["Bouquets"],
        "summary": "Get a bouquet by id",
        "parameters": [{ "$ref": "#/components/parameters/BouquetId" }],
        "responses": {
          "200": {
            "description": "Bouquet found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Bouquet" }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "tags": ["Bouquets"],
        "summary": "Update a bouquet by id",
        "parameters": [{ "$ref": "#/components/parameters/BouquetId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BouquetUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bouquet updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Bouquet" }
              }
            }
          },
          "400": { "description": "Empty or invalid body" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "tags": ["Bouquets"],
        "summary": "Delete a bouquet by id",
        "parameters": [{ "$ref": "#/components/parameters/BouquetId" }],
        "responses": {
          "200": {
            "description": "Bouquet deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Bouquet" }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "BouquetId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "schemas": {
      "Bouquet": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "photoURL": { "type": "string" },
          "favorite": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "BouquetUpdate": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "photoURL": { "type": "string" },
          "favorite": { "type": "boolean" }
        }
      }
    }
  }
}